backend/requirements.txt
crewai
python-dotenv
unstructured
langchain_community
langchain_groq
langchain_openai
fastapi
uvicorn
python-multipart
beautifulsoup4
lxml
requests
sec-api
langchain-mistralai
langchain-google-genai
langchain-openrouter
httpx

PyPDF2

